workers/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
